Add CMake support
Providing:
- target (library) `SetReplace`
- config files for third-parties CMake projects
to use SetReplace with no need of FindSetReplace.cmake
- Usable from build and install trees.
- Option to use with or without Mathematica.
- Infrastructure to perform cpp tests with gtest.

CMake options:

- `SET_REPLACE_BUILD_TESTING`:
Download gtest at build time (with FetchContent) and enable
basic cpp testing.

- `SET_REPLACE_WITH_MATHEMATICA`:
Compiles `SetReplace.cpp` which provides the interface for Wolfram Language.
For this `FindMathematica.cmake` is needed.
`FindMathematica.cmake` is taken from:
https://github.com/sakra/FindMathematica/blob/master/CMake/Mathematica/FindMathematica.cmake
that seems used and up-to-date.
Please note that FindX.cmake is deprecated in CMake, and CONFIG files should be
provided by the project (Mathematica). But this is an upstream issue,
out of our control.

Closes #339